requirements.txt
# Core Libraries
streamlit
pandas
torch
gc
torchaudio
whisper
pyarrow

# LangChain & AI Models
langchain
langchain_community
langchain-google-genai
sentence-transformers
chromadb

# Google Gemini API
google-generativeai
google-auth
google-auth-oauthlib
google-auth-httplib2

# Environment Variables
python-dotenv